+++ lib/modules/radarr/modules/add_movie/widgets/page_search.dart
@@ -7,19 +7,18 @@ class RadarrAddMovieSearchPage extends StatefulWidget {
     State<StatefulWidget> createState() => _State();
 }
 
-class _State extends State<RadarrAddMovieSearchPage> with AutomaticKeepAliveClientMixin, LunaLoadCallbackMixin {
+class _State extends State<RadarrAddMovieSearchPage> with AutomaticKeepAliveClientMixin {
     final GlobalKey<ScaffoldState> _scaffoldKey = GlobalKey<ScaffoldState>();
     final GlobalKey<RefreshIndicatorState> _refreshKey = GlobalKey<RefreshIndicatorState>();
 
     @override
     bool get wantKeepAlive => true;
 
-    @override
     Future<void> loadCallback() async {
         if(context.read<RadarrAddMovieState>().searchQuery.isNotEmpty) {
-            context.read<RadarrAddMovieState>().executeSearch(context);
-            await context.read<RadarrState>()?.lookup;
-            await context.read<RadarrAddMovieState>()?.exclusions;
+            context.read<RadarrAddMovieState>().fetchLookup(context);
+            await context.read<RadarrAddMovieState>().lookup;
+            await context.read<RadarrAddMovieState>().exclusions;
         }
     }
 
@@ -43,15 +42,12 @@ class _State extends State<RadarrAddMovieSearchPage> with AutomaticKeepAliveClie
     Widget _body() {
         return Selector<RadarrState, Future<List<RadarrMovie>>>(
             selector: (_, state) => state.movies,
-            builder: (context, movies, _) => Selector<RadarrAddMovieState, Future<List<RadarrExclusion>>>(
-                selector: (_, state) => state.exclusions,
-                builder: (context, exclusions, _) => Selector<RadarrState, Future<List<RadarrMovie>>>(
-                    selector: (_, state) => state.lookup,
-                    builder: (context, lookup, _) {
-                        if(lookup == null) return Container();
-                        return _builder(movies, lookup, exclusions);
-                    },
-                ),
+            builder: (context, movies, _) => Selector<RadarrAddMovieState, Tuple2<Future<List<RadarrMovie>>, Future<List<RadarrExclusion>>>>(
+                selector: (_, state) => Tuple2(state.lookup, state.exclusions),
+                builder: (context, tuple, _) {
+                    if(tuple.item1 == null) return Container();
+                    return _builder(movies, tuple.item1, tuple.item2);
+                },
             ),
         );
     }
@@ -72,24 +68,25 @@ class _State extends State<RadarrAddMovieSearchPage> with AutomaticKeepAliveClie
                         );
                         return LunaMessage.error(onTap: _refreshKey.currentState.show);
                     }
-                    if(snapshot.hasData) return _results(snapshot.data[0], snapshot.data[1], snapshot.data[2]);
+                    if(snapshot.hasData) return _list(snapshot.data[0], snapshot.data[1], snapshot.data[2]);
                     return LunaLoader();
                 },
             ),
         );
     }
 
-    Widget _results(List<RadarrMovie> movies, List<RadarrMovie> results, List<RadarrExclusion> exclusions) {
+    Widget _list(List<RadarrMovie> movies, List<RadarrMovie> results, List<RadarrExclusion> exclusions) {
         if((results?.length ?? 0) == 0) return LunaListView(children: [LunaMessage.inList(text: 'No Results Found')]);
         return LunaListViewBuilder(
             scrollController: RadarrAddMovieNavigationBar.scrollControllers[0],
             itemCount: results.length,
             itemBuilder: (context, index) {
                 RadarrExclusion exclusion = exclusions?.firstWhere((exclusion) => exclusion.tmdbId == results[index].tmdbId, orElse: () => null);
+                RadarrMovie movie = movies?.firstWhere((movie) => movie.id == results[index].id, orElse: () => null);
                 return RadarrAddMovieSearchResultTile(
                     movie: results[index],
-                    exists: results[index].id != null,
-                    isExcluded: !(exclusion == null),
+                    exists: movie != null,
+                    isExcluded: exclusion != null,
                 );
             },
         );