+++ lib/modules/radarr/modules/add_movie/widgets/search_search_bar.dart
@@ -32,7 +32,7 @@ class _State extends State<RadarrAddMovieSearchSearchBar> {
                                 autofocus: context.read<RadarrAddMovieState>().searchQuery?.isEmpty ?? true,
                                 onChanged: (value) => context.read<RadarrAddMovieState>().searchQuery = value,
                                 onSubmitted: (value) {
-                                    if(value.isNotEmpty) context.read<RadarrAddMovieState>().executeSearch(context);
+                                    if(value.isNotEmpty) context.read<RadarrAddMovieState>().fetchLookup(context);
                                 },
                                 margin: EdgeInsets.fromLTRB(12.0, 0.0, 12.0, 14.0),
                             ),